+++ src/pages/UserLookbookCreate/LookbookCreateHeader/LookbookCreateHeader.tsx
@@ -1,7 +1,7 @@
 import GoBackButton from '@/components/GoBackButton/GoBackButton';
 import Header from '@/components/Header/Header';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { C, S } from './style';
+import { C, S } from './LookbookCreateHeader.style';
 import { LocationState, LookbookSelect } from '../UserLookbookCreatePage';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { WeatherType } from '@/types/weather';
@@ -9,7 +9,7 @@ import { saveLookbook } from '@/services/clothes';
 import useAppSelector from '@/hooks/useAppSelector';
 import { useSnackbar } from '@/contexts/SnackbarProvider';
 import { useState } from 'react';
-import DeleteDialog from './DeleteDialog';
+import DeleteDialog from './DeleteDialog/DeleteDialog';
 
 type LookbookCreateHeaderProps = {
   weatherType: WeatherType;
