+++ src/pages/UserLookbookCreate/UserLookbookCreatePage.tsx
@@ -1,5 +1,4 @@
 import { Navigate, useLocation, useSearchParams } from 'react-router-dom';
-import LookbookCreateHeader from './LookbookCreateHeader';
 import { S } from './UserLookbookCreatePage.style';
 import WeatherHeadline from './WeatherHeadline/WeatherHeadline';
 import { WeatherType } from '@/types/weather';
@@ -9,6 +8,7 @@ import { Outfits } from '@/types/clothes';
 import { TempCondition } from '../Home/ClothesSection/ClothesSection';
 import HeadHelmet from '@/components/HeadHelmet/HeadHelmet';
 import { DEFAULT_CLOTHES_BY_WEATHER } from '@/constants/lookbook';
+import LookbookCreateHeader from './LookbookCreateHeader/LookbookCreateHeader';
 
 export type LookbookSelect = {
   top: { name: string; color: string };
